added functionalities

+++ My-Project/src/main/java/com/ezinne/note/NoteService.java
@@ -6,8 +6,11 @@ import com.ezinne.appUser.AppUserService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.time.ZonedDateTime;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @Service
@@ -19,10 +22,11 @@ public class NoteService {
 
     public String saveNotes(Long userId, Note notes) {
 
-        AppUser appUser = appUserRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("user not found"));
-
-
+        Optional<AppUser> optionalAppUser = appUserRepository.findById(userId);
+        if (optionalAppUser.isEmpty()) {
+            return "user not found";
+        }
+        AppUser appUser = optionalAppUser.get();
         Boolean loggedIn = appUser.getLoggedIn();
         if (loggedIn) {
 
@@ -30,19 +34,22 @@ public class NoteService {
         notes.setAppUser(appUser);
         noteRepository.save(notes);
 
-        return "Title - " + notes.getNoteTitle() + ": " + notes.getMyNotes() + " created at " + notes.getTimestamp();
+        return "Title - " + notes.getNoteTitle() + ": " + notes.getMyNotes() + " with id: " + notes.getId() + " created at " + notes.getTimestamp();
     }
-        return"user needs to login";
+        return "user needs to login";
 
     }
     public String editNotes(Long noteId, Note newNote) {
-        Note existingNotes = noteRepository.findById(noteId).
-                    orElseThrow(() -> new IllegalArgumentException(String.format("%d does not exist", noteId)));
+        Optional<Note> optionalNote = noteRepository.findById(noteId);
+        if (optionalNote.isEmpty()){
+            return String.format("%d does not exist", noteId);
+        }
+        Note note = optionalNote.get();
 
-        existingNotes.setMyNotes(newNote.getMyNotes());
-        newNote.setId(existingNotes.getId());
-        newNote.setNoteTitle(existingNotes.getNoteTitle());
-        newNote.setAppUser(existingNotes.getAppUser());
+        note.setMyNotes(newNote.getMyNotes());
+        newNote.setId(note.getId());
+        newNote.setNoteTitle(note.getNoteTitle());
+        newNote.setAppUser(note.getAppUser());
         newNote.setTimestamp(ZonedDateTime.now());
 
         noteRepository.save(newNote);
@@ -50,16 +57,66 @@ public class NoteService {
 
     }
 
-    public List<Note> listOfNotesForUser(Long appUserId) {
+    public String listOfNotesForUser(Long appUserId) {
 
-        AppUser appUser = appUserRepository.findById(appUserId)
-                .orElseThrow(() -> new IllegalArgumentException("user not found"));
+        Optional<AppUser> optionalAppUser = appUserRepository.findById(appUserId);
+        if (optionalAppUser.isEmpty()) {
+            return "user not found";
+        }
+        AppUser appUser = optionalAppUser.get();
+        List<Note> notes = appUser.getNotes();
+        StringBuilder sb = new StringBuilder();
+        for (Note note : notes) {
+            sb.append(note.getId());
+            sb.append(" - ");
+            sb.append(note.getNoteTitle());
+            sb.append(": ");
+            sb.append(note.getMyNotes());
+            sb.append("\n");
+        }
+       return sb.toString();
 
-        return appUser.getNotes();
+    }
+
+    public String notesByPeriodForUser(Long appUserId, LocalDate start, LocalDate end) {
+        Optional<AppUser> optionalAppUser = appUserRepository.findById(appUserId);
+        if (optionalAppUser.isEmpty()) {
+            return "user not found";
+        }
+
+        AppUser appUser = optionalAppUser.get();
+        List<Note> notes = appUser.getNotes();
+
+        StringBuilder sb = new StringBuilder();
+        if (notes.isEmpty()) {
+            return "No notes was found for this user";
+        }
+        for (Note note : notes) {
+            ZonedDateTime timestamp = note.getTimestamp();
+            LocalDate date = timestamp.toLocalDate();
+            if (date.isEqual(start) || date.isEqual(end) || (date.isAfter(start) && date.isBefore(end))) {
+                sb.append(note.getId());
+                sb.append(" - ");
+                sb.append(note.getNoteTitle());
+                sb.append(": ");
+                sb.append(note.getMyNotes());
+                sb.append(": ");
+                sb.append(note.getTimestamp());
+                sb.append("\n");
+                continue;
+            }
+            return "No notes was created within this period";
+        }
+
+        return sb.toString();
 
     }
 
     public String deleteNoteById(Long noteId) {
+        Optional<Note> noteOptional = noteRepository.findById(noteId);
+        if (noteOptional.isEmpty()) {
+            return String.format("note with %d does not exist", noteId);
+        }
         noteRepository.deleteById(noteId);
         return String.format("note with %d has been deleted", noteId);
     }
